Feat : 서류 관련 API 구현

+++ backend/main-service/src/main/java/com/turtlecoin/mainservice/domain/turtle/dto/AuctionTurtleInfoDTO.java
@@ -9,5 +9,5 @@ import lombok.Data;
 public class AuctionTurtleInfoDTO {
     private Long id;
     private Gender gender;
-    private Double size;
+    private int size;
 }
